+++ src/Symfony/Component/Form/Tests/Extension/Core/ChoiceList/ObjectChoiceListTest.php
@@ -39,6 +39,9 @@ class ObjectChoiceListTest extends \PHPUnit_Framework_TestCase
 
     private $obj4;
 
+    /**
+     * @var ObjectChoiceList
+     */
     private $list;
 
     protected function setUp()